requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.3
tqdm==4.66.1
census==0.8.19
us==3.1.0
geopandas==0.14.2
shapely==2.0.2
folium==0.15.1
geopy==2.4.1
vaderSentiment==3.3.2
scikit-learn==1.4.0
xgboost==2.0.3
shap==0.44.0
imbalanced-learn==0.11.0
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
jupyterlab==4.0.10
ipykernel==6.29.0